Corrects CourseMarks model

+++ db/migrate/20200730091307_create_course_marks.rb
@@ -2,7 +2,7 @@ class CreateCourseMarks < ActiveRecord::Migration[6.0]
   def change
     create_table :course_marks do |t|
       t.references :student_course, null: false, foreign_key: true
-      t.numeric :mark, null: false, default: 10
+      t.float :mark, null: false, default: 10
       t.integer :attendance, limit: 2, null: false, default: 0
       t.text :observations
